refactor: extract parseLevel and updateLevels into new file
